Re-organizing short-range code post-merge, removing old references to refactored type
