+++ src/wr_logic_ai/src/wr_logic_ai/shortrange/encoder_drive.py
@@ -7,7 +7,7 @@ import rospy
 from std_msgs.msg import Float64
 
 from shortrange_util import ShortrangeStateEnum, ShortrangeState
-from wr_logic_ai.src.shortrange_util import ShortrangeStateEnum
+from wr_logic_ai.shortrange.shortrange_util import ShortrangeStateEnum
 from wrevolution.srv import ResetEncoder
 
 # Base speed for robot to move at
@@ -32,6 +32,7 @@ right_motor_pub = rospy.Publisher(right_motor_topic, Float64)
 
 
 def distance_to_encoder(meters: float) -> int:
+    # TODO (@bennowotny @co-li ) Math here is suspicious
     return int(ENCODER_COUNTS_PER_ROTATION * meters / (ROBOT_WHEEL_DIAMETER / 2 * math.pi))
 
 
@@ -47,7 +48,7 @@ class EncoderDrive(ShortrangeState):
 
     def encoder_callback(self, motor_pub: rospy.Publisher, setpoint: int, done_idx: int, encoder: Float64):
         if (encoder.data < setpoint):
-            motor_pub.publish(Float64(SPEED)) 
+            motor_pub.publish(Float64(SPEED))
         else:
             self.is_done[done_idx] = True
             motor_pub.publish(Float64(0))
@@ -57,12 +58,14 @@ class EncoderDrive(ShortrangeState):
 
         self.reset_encoder(left_encoder_reset_srv)
         self.reset_encoder(right_encoder_reset_srv)
-        sub_left = rospy.Subscriber(left_encoder_topic, Float64, lambda msg : self.encoder_callback(left_motor_pub, self.setpoint, 0, msg))
-        sub_right = rospy.Subscriber(right_encoder_topic, Float64, lambda msg: self.encoder_callback(right_motor_pub, self.setpoint, 1, msg))
+        sub_left = rospy.Subscriber(left_encoder_topic, Float64, lambda msg: self.encoder_callback(
+            left_motor_pub, self.setpoint, 0, msg))
+        sub_right = rospy.Subscriber(right_encoder_topic, Float64, lambda msg: self.encoder_callback(
+            right_motor_pub, self.setpoint, 1, msg))
 
         while not all(self.is_done):
             rospy.sleep(rate)
-        
+
         sub_left.unregister()
         sub_right.unregister()
